1. Calificaciones de gastronomia por documento usuario

+++ src/main/java/com/project/dto/UsuarioTurismoDTO.java
@@ -13,7 +13,7 @@ public class UsuarioTurismoDTO implements Serializable {
 	private static final long serialVersionUID = 1L;
 
 	private Integer ustucodi;
-	private Integer documento;
+	private Long documento;
 	private Integer turicodi;
 	private Integer like;
 }
